alexa/requirements.txt
ask-alexa-pykit=0.5.6
requests
